Avoid out of order mounting and updates of the Canvas component
Because these hooks are asynchronous, they were not strictly ordered. This
caused updates to be rendered before mounts and certain updates out of order.
The new "queue" (with no buffer, only 1 active message) ensures that only the
last received update is rendered at any time. This mechanism also ensures that
no updates are processed before the component is properly mounted and
initialized.